+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [2.1.0] - 2026-08-05
+
+### Changed
+
+- Generated mirrors are declared in the namespace of the script they mirror.
+- `Generate Controllable Script` moved from `Assets ▸ Controllable` to `Assets ▸ OCF`.
+
 ## [2.0.0] - 2026-07-22
 
 ### Added